Started a debugger UI

+++ language/system/OtOS.cpp
@@ -271,7 +271,7 @@ OtObject OtOSClass::networks() {
 //
 
 double OtOSClass::clock() {
-	return static_cast<float>(std::chrono::duration_cast<std::chrono::milliseconds>(
+	return static_cast<double>(std::chrono::duration_cast<std::chrono::milliseconds>(
 		std::chrono::system_clock::now().time_since_epoch()).count()) / 1000.0;
 }
